car_server/requirements.txt
pyzmq==15.2.0
pyserial==3.0.1